requirements.txt
# python packages
tensorflow2
PyAudio (sudo apt install python3-pyaudio)
numpy
matplotlib
audiofile

# other software:
OpenSmile
